requirements.txt
gmplot==1.4.1
ipython==8.12.3
joblib==1.4.2
matplotlib==3.9.2
numpy==2.1.2
pandas==2.2.3
PyYAML==6.0.2
PyYAML==6.0.2
scikit_learn==1.5.2
scipy==1.14.1
